@@ -25,8 +25,8 @@
     "devicons-react": "1.3.0-beta.0.1.4",
     "framer-motion": "^10.18.0",
     "lucide-react": "^0.395.0",
-    "next": "14.2.5",
-    "next-intl": "^3.15.2",
+    "next": "14.2.14",
+    "next-intl": "^3.20.0",
     "next-themes": "^0.3.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
@@ -45,9 +45,9 @@
     "@types/react-dom": "18.0.11",
     "autoprefixer": "^10.4.19",
     "eslint": "8.36.0",
-    "eslint-config-next": "13.2.4",
+    "eslint-config-next": "14.2.14",
     "postcss": "^8.4.38",
     "tailwindcss": "^3.4.4",
-    "typescript": "4.9.5"
+    "typescript": "5.6.2"
   }
 }
