@@ -1,4 +1,5 @@
-const createNextIntlPlugin = require('next-intl/plugin');
+import createNextIntlPlugin from 'next-intl/plugin';
+
 const withNextIntl = createNextIntlPlugin();
 
 /** @type {import('next').NextConfig} */
@@ -17,4 +18,4 @@ const nextConfig = {
   }
 }
 
-module.exports = withNextIntl(nextConfig)
+export default withNextIntl(nextConfig)
